backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

python-chess==1.999

langchain
langchain-openai
openai

pydantic==2.9.2
pydantic-settings==2.5.2

pytest==8.3.3
httpx==0.27.2